backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pytest==7.4.0
httpx==0.24.1  # Required for TestClient
fastapi-mail==1.4.1
python-magic==0.4.27  # for file type detection
aiofiles==23.2.1     # for async file operations
pillow==10.1.0       # for image processing
email-validator==2.1.0.post1
